pyproject.toml
[project]
name = "eu-ai-risks"
version = "0.1.0"
description = "Parse the EU AI Act into a Neo4j graph database."
requires-python = ">=3.12"
dependencies = [
    "neo4j",
    "pdfplumber",
    "python-dotenv",
]

[project.scripts]
eu-ai-risks = "eu_ai_risks.main:main"
